requirements.txt
QDarkStyle==2.5.4
setuptools==39.1.0
PyOpenGL==3.1.0
